test(reader): split reader chrome coverage
Why:
- Continue the test architecture migration by moving pure BibleUI reader chrome coverage out of the app-host XCTest bundle.
- Preserve Android-style reader header, toolbar, popup, overflow, Downloads routing, and keyboard shortcut contracts in the package lane.

What Changed:
- Added ReaderChromeTests under the BibleUI package test target.
- Removed the migrated reader chrome tests from AndBibleTests+AppAndReader.
- Cleaned stale SwiftUI imports left behind by the migration.
- Updated the migration ledger for the reader chrome slice.

Validation:
- git diff --check
- python3 scripts/check_repo_standards.py docblocks --all-files
- python3 scripts/check_repo_standards.py source-guards
- xcodebuild test -scheme BibleUITests -only-testing:BibleUITests/ReaderChromeTests
- xcodebuild test -scheme AndBible -only-testing:AndBibleTests/AndBibleTests/testModuleBrowserFiltersAndSortsAndroidDownloadRows
- GitNexus impact reviewed for the reader chrome symbols before editing.
- GitNexus detect_changes reported low risk with no affected execution flows.
- Adversarial review completed; stale import feedback was addressed.

Impact:
- Reduces app-host unit-test load while preserving reader chrome behavior coverage.
- No production behavior changes.
