+++ docs/test-architecture-migration-plan.md
@@ -156,6 +156,7 @@ Blocker: `AndBibleTestSupport.swift` is a 2,233-line `extension AndBibleTests` e
 - `AndBibleTests+AppAndReader` text-display slice has been split by owning module: app-default, inheritance, full-resolution, and dirty-override cleanup contracts moved to a dedicated `BibleCoreTests` suite, while Strong's legacy-mode bridge normalization moved to `BibleUITests`.
 - `AndBibleTests+AppAndReader` SWORD coordinator slice has moved to `BibleUITests` because it exercises BibleUI reader SWORD setup, installed-module categorization, active-module fallback, book-list derivation, and display-option application without app delegate/bootstrap behavior.
 - `AndBibleTests+AppAndReader` module-switch planner slice has moved to `BibleUITests` because it exercises BibleUI current-document switch planning, Android category mismatch rejection, atomic auxiliary-document persistence, module-only selection, and category reload decisions without SWORD fixtures or app bootstrap.
+- `AndBibleTests+AppAndReader` reader chrome slice has moved to `BibleUITests` because it exercises BibleUI header, toolbar action, popup placement, overflow menu, Downloads routing, and keyboard shortcut construction without app delegate/bootstrap behavior.
 - Batches for the genuinely BibleUI-behavior remainder: ReaderNavigation, Bookmarks/Strongs, Window/Settings, view-model parts of `+AppAndReader`.
 - **Split `+AppAndReader` deliberately**: `sceneConfiguration` test stays app-hosted; the `ContentView` legacy root-sidebar source scan is now a repo-standards `source-guards` check; the remaining package-owned tests should move to the lowest owning package target.
 - Run BibleUITests via `xcodebuild test -scheme BibleUITests -destination 'platform=iOS Simulator,...'` against the **committed package test scheme from Phase 0** (no app host).
